+++ Service/PostEntity.php
@@ -25,4 +25,25 @@ final class PostEntity extends VirtualEntity
     {
         return date($format, $this->getTimestamp());
     }
+
+    /**
+     * Returns image URL
+     * 
+     * @param string $size
+     * @return string
+     */
+    public function getImageUrl($size)
+    {
+        return $this->getImageBag()->getUrl($size);
+    }
+
+    /**
+     * Determines whether entity has a cover image
+     * 
+     * @return boolean
+     */
+    public function hasCover()
+    {
+        return $this->getCover() !== '';
+    }
 }